feat: add docker inside the vm

+++ justfile
@@ -18,8 +18,8 @@ ssh:
 # Log in to OpenFaaS
 faas-login:
     #!/usr/bin/env bash
-    PASS=$($SSH_CMD sudo kubectl get secret -n openfaas basic-auth -o jsonpath="{.data.basic-auth-password}" | base64 --decode)
-    echo $PASS | faas-cli login --password-stdin --gateway http://127.0.0.1:31112
+    PASS=$({{SSH_CMD}} sudo kubectl get secret -n openfaas basic-auth -o jsonpath="{.data.basic-auth-password}" | base64 --decode)
+    echo $PASS | {{SSH_CMD}} faas-cli login --password-stdin --gateway http://127.0.0.1:31112
     echo "PASSWORD: $PASS"
 
 # Publish OpenFaaS functions using the registry variables and modify the function file
@@ -37,26 +37,30 @@ faas-pub-single file:
     #!/usr/bin/env bash
     cd {{ justfile_directory() }}/functions
     sed -i "s|image: .*|image: {{ REGISTRY }}/$(basename {{file}} .yml):{{ TAG }}|" "{{file}}"
+    {{SSH_CMD}} << EOF
+    cd /home/myce/mycelium/functions
     faas-cli publish -f "{{file}}"
     faas-cli deploy -f "{{file}}"
+    EOF
 
 # Publish a test message to MQTT broker (run this command from within the VM)
 mqtt:
+    #!/usr/bin/env bash
     {{SSH_CMD}} 'mosquitto_pub -h localhost -t sample-topic -m "Hello World!"'
 
 # Build and run the VM
-vm:
+vm clean='false':
     #!/usr/bin/env bash
     set -e
     SYSTEM=$(case "$(uname -s)-$(uname -m)" in \
         Linux-x86_64)  echo "x86_64-linux" ;; \
-        Darwin-arm64)  echo "aarch64-linux" ;; \
-        Darwin-x86_64) echo "x86_64-linux" ;; \
+        Linux-aarch64)  echo "aarch64-linux" ;; \
+        Darwin-x86_64) echo "x86_64-darwin" ;; \
         *)             echo "Unsupported system" && exit 1 ;; \
     esac)
     HOST_SYSTEM=$(case "$(uname -s)-$(uname -m)" in \
         Linux-x86_64)  echo "x86_64-linux" ;; \
-        Darwin-arm64)  echo "aarch64-darwin" ;; \
+        Linux-aarch64)  echo "aarch64-linux" ;; \
         Darwin-x86_64) echo "x86_64-darwin" ;; \
         *)             echo "Unsupported system" && exit 1 ;; \
     esac)
@@ -68,7 +72,9 @@ vm:
       install = vm.config.system.build.vm;
     in
     install" )
+    if [ "{{clean}}" = "true" ]; then
     rm *.qcow2 || true
+    fi
     if [[ "$(uname -s)" == "Linux" ]]; then
     exec $vmpath/bin/run-* -nographic -cpu host -enable-kvm
     else
